Create visual_narrative
The narrative visualisation explores the crime rate of suburbs in Melbourne, normalised by their residential populations. Broadly, geographical areas of suburbs are compared and contrasted with one another to give an idea of which areas have more crime. The message then focuses on individual suburbs with higher crime rates and/or high change in crime rate and highlights the trends of those suburbs over the past decade. The intended audience is anyone with a general interest in the data and could include prospective property buyers or renters in Melbourne.